api/requirements.txt
amqp==2.6.0
asgiref==3.2.7
autopep8==1.5.2
billiard==3.6.3.0
celery==4.4.6
click==7.1.2
coreapi==2.3.3
croniter==0.3.31
Django==3.0.7
django-admin-autocomplete-filter==0.5
django-admin-list-filter-dropdown==1.0.3
django-apptemplates==1.5
django-celery-results==1.2.1
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-decorator-include==2.1
django-invitations==1.9.3
django-inlinecss==0.3.0
django-model-utils==4.0.0
djangorestframework==3.12.2
django-rq==2.3.2
django-sitetree==1.14.0
django-werkzeug==1.0.0
future==0.18.2
gunicorn==20.0.4
importlib-metadata==1.6.0
kombu==4.6.11
Markdown==3.2.1
mccabe==0.6.1
model-bakery==1.2.1
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.4.1
redis==3.5.2
rq==1.4.1
rq-scheduler==0.10.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
vine==1.3.0
whitenoise==5.0.1
zipp==3.1.0
